centralized authorization for all API, cleaned the code removed unused variables and imported

+++ src/main/java/com/satwik/splitwiseclone/service/implementations/GroupServiceImpl.java
@@ -15,59 +15,53 @@ import com.satwik.splitwiseclone.repository.UserRepository;
 import com.satwik.splitwiseclone.service.interfaces.GroupService;
 import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
-import java.nio.file.AccessDeniedException;
 import java.util.*;
 
 @Service
 public class GroupServiceImpl implements GroupService {
 
     @Autowired
-    GroupRepository groupRepository;
+    private AuthorizationService authorizationService;
 
     @Autowired
-    UserRepository userRepository;
+    private GroupRepository groupRepository;
 
     @Autowired
-    ExpenseRepository expenseRepository;
+    private UserRepository userRepository;
 
     @Autowired
-    GroupMembersRepository groupMembersRepository;
+    private ExpenseRepository expenseRepository;
+
+    @Autowired
+    private GroupMembersRepository groupMembersRepository;
+
 
     @Override
     @Transactional
-    public String createGroup(GroupDTO groupDTO, UUID userId) throws AccessDeniedException {
+    public String createGroup(GroupDTO groupDTO) {
 
-        User user = userRepository.findById(userId).orElseThrow(() -> new RuntimeException("User not found"));
+        User user = authorizationService.getAuthorizedUser();
 
-        // create the group
         Group group = new Group();
         group.setGroupName(groupDTO.getGroupName());
         group.setUser(user);
         group.setDefaultGroup(false);
-
-        // save the group
         groupRepository.save(group);
 
         return "Group created successfully!";
     }
 
     @Override
-    public GroupListDTO findAllGroup(UUID userId) throws AccessDeniedException {
+    public GroupListDTO findAllGroup() {
 
-        User user = userRepository.findById(userId).orElseThrow(() -> new RuntimeException("User not found"));
-
+        User user = authorizationService.getAuthorizedUser();
         List<Group> groupList = groupRepository.findByUserId(user.getId());
-
         GroupListDTO groupListDTO = new GroupListDTO();
         groupListDTO.setOwner(user.getUsername());
 
         List<GroupListDTOWithin> groupListDTOS = new ArrayList<>();
-
-        // mapping each group list to group dto list
         for (Group group : groupList) {
             groupListDTOS.add(new GroupListDTOWithin(group.getId(), group.getGroupName()));
         }
@@ -78,7 +72,9 @@ public class GroupServiceImpl implements GroupService {
 
     @Transactional
     @Override
-    public String addGroupMembers(UUID groupId, UUID memberId) {
+    public String addGroupMembers(UUID groupId, UUID memberId) throws Exception {
+
+        authorizationService.checkAuthorizationOnGroup(groupId);
 
         Group group = groupRepository.findById(groupId).orElseThrow(() -> new RuntimeException("Group not found!"));
         User member = userRepository.findById(memberId).orElseThrow(() -> new RuntimeException("User not found to add as member!"));
@@ -91,11 +87,12 @@ public class GroupServiceImpl implements GroupService {
     }
 
     @Override
-    public List<UserDTO> findMembers(UUID groupId) {
+    public List<UserDTO> findMembers(UUID groupId) throws Exception {
+
+        authorizationService.checkAuthorizationOnGroup(groupId);
 
         Group group = groupRepository.findById(groupId).orElseThrow(() -> new RuntimeException("Group not found!"));
         List<GroupMembers> groupMembersList = groupMembersRepository.findByGroupId(group.getId());
-        // mapping to userdto
         List<UserDTO> userDTOS = new ArrayList<>();
         for (GroupMembers groupMembers : groupMembersList) {
             UserDTO userDTO = new UserDTO();
@@ -104,13 +101,15 @@ public class GroupServiceImpl implements GroupService {
             userDTO.setPhone(new PhoneDTO(groupMembers.getMember().getCountryCode(), groupMembers.getMember().getPhoneNumber()));
             userDTOS.add(userDTO);
         }
+
         return userDTOS;
     }
 
     @Transactional
     @Override
-    public String deleteMembers(UUID groupMemberId) {
+    public String deleteMembers(UUID groupId, UUID groupMemberId) throws Exception {
 
+        authorizationService.checkAuthorizationOnGroup(groupId);
 
         groupMembersRepository.deleteById(groupMemberId);
 
@@ -119,39 +118,28 @@ public class GroupServiceImpl implements GroupService {
 
     @Override
     @Transactional
-    public String deleteGroupByGroupId(UUID groupId, UUID userId) throws AccessDeniedException {
-
-        User user = userRepository.findById(userId).orElseThrow(() -> new RuntimeException("User not found"));
-        Group group = groupRepository.findById(groupId).orElseThrow(() -> new RuntimeException("Group not found"));
-
-        if (user == null || user.getId() != group.getUser().getId()) throw new AccessDeniedException("Access Denied");
-
+    public String deleteGroupByGroupId(UUID groupId) throws Exception {
+        
+        authorizationService.checkAuthorizationOnGroup(groupId);
+        
         // TODO : add code to check the default group (default group can't be deleted)
         groupRepository.deleteById(groupId);
-        return "Successfully deleted the group - %d.".formatted(groupId);
+        return "Successfully deleted the group - %s.".formatted(groupId);
 
     }
 
     @Override
-    public GroupDTO findGroupByGroupId(UUID groupId, UUID userId) throws AccessDeniedException {
+    public GroupDTO findGroupByGroupId(UUID groupId) throws Exception {
 
-        User user = userRepository.findById(userId).orElseThrow(() -> new RuntimeException("User not found"));
-        Group group = groupRepository.findById(groupId).orElseThrow(() -> new RuntimeException("Group not found"));
-
-        if(user == null || user.getId() != group.getUser().getId()) throw new AccessDeniedException("Access Denied");
-
+        Group group = authorizationService.checkAuthorizationOnGroup(groupId);
         GroupDTO groupDTO = new GroupDTO();
-
         groupDTO.setGroupId(group.getId());
         groupDTO.setGroupName(group.getGroupName());
         groupDTO.setOwner(group.getUser().getUsername());
 
         List<Expense> expenseList = expenseRepository.findByGroupId(groupId);
         List<ExpenseListDTO> expenseDTOList = new ArrayList<>();
-
         for (Expense expense : expenseList) {
-
-            // mapping expense to expense DTO
             ExpenseListDTO expenseListDTO = new ExpenseListDTO();
             expenseListDTO.setAmount(expense.getAmount());
             expenseListDTO.setDescription(expense.getDescription());
@@ -160,12 +148,17 @@ public class GroupServiceImpl implements GroupService {
         }
         groupDTO.setExpenses(expenseDTOList);
 
+        List<GroupMemberDTO> groupMemberDTOS = getGroupMemberDTOS(group);
+        groupDTO.setGroupMembers(groupMemberDTOS);
+
+        return groupDTO;
+    }
+
+    private static List<GroupMemberDTO> getGroupMemberDTOS(Group group) {
         List<GroupMembers> groupMembers = group.getGroupMembers();
         List<GroupMemberDTO> groupMemberDTOS = new ArrayList<>();
-
         for (GroupMembers groupMember : groupMembers) {
             User user1 = groupMember.getMember();
-            // map user to group member dto
             GroupMemberDTO groupMemberDTO = new GroupMemberDTO();
             groupMemberDTO.setGroupMemberId(groupMember.getId());
             groupMemberDTO.setMemberId(user1.getId());
@@ -173,23 +166,19 @@ public class GroupServiceImpl implements GroupService {
             groupMemberDTO.setUsername(user1.getUsername());
             groupMemberDTOS.add(groupMemberDTO);
         }
-        groupDTO.setGroupMembers(groupMemberDTOS);
-
-        return groupDTO;
+        return groupMemberDTOS;
     }
 
     @Override
     @Transactional
-    public String updateGroup(GroupUpdateRequest groupUpdateRequest, UUID groupId, UUID userId) throws AccessDeniedException {
+    public String updateGroup(GroupUpdateRequest groupUpdateRequest, UUID groupId) throws Exception {
 
-        User user = userRepository.findById(userId).orElseThrow(() -> new RuntimeException("User not found"));
+        authorizationService.checkAuthorizationOnGroup(groupId);
+        
         Group group = groupRepository.findById(groupId).orElseThrow(() -> new RuntimeException("Group not found"));
-
-        if(user == null || user.getId() != group.getUser().getId()) throw new AccessDeniedException("Access Denied");
-
         group.setGroupName(groupUpdateRequest.getGroupName());
         groupRepository.save(group);
 
-        return "%d - Group update successfully!".formatted(group.getId());
+        return "%s - Group update successfully!".formatted(group.getId());
     }
 }
